pyproject.toml
[project]
name = "meta-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "maxbrzr", email = "maximilian.burzer@kit.edu" }
]
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "mlflow>=3.9.0",
    "ncps>=1.0.1",
    "seaborn>=0.13.2",
    "torch>=2.10.0",
    "wandb>=0.25.1",
    "whar-datasets",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
whar-datasets = { git = "https://github.com/teco-kit/whar-datasets.git" }
